Implementação do AutorValidator e da validação do cadastro de livros vinculado ao autor

+++ Libraryapi/src/main/java/com/rodrigogustavo/Libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package com.rodrigogustavo.Libraryapi.service;
 
-import com.rodrigogustavo.Libraryapi.controller.AutorDTO;
 import com.rodrigogustavo.Libraryapi.domain.Autor;
+import com.rodrigogustavo.Libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.rodrigogustavo.Libraryapi.repositories.AutorRepository;
+import com.rodrigogustavo.Libraryapi.repositories.LivroRepository;
+import com.rodrigogustavo.Libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,23 +13,27 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor //Cria um contrutor com os campos obrigatórios
 public class AutorService {
 
     private final AutorRepository repository;
-
-    public AutorService(AutorRepository repository){
-        this.repository = repository;
-    }
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
     public Optional<Autor> obterDetalhes(UUID id){
+
         return repository.findById(id);
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido exclui auto que possui livro cadastrado");
+        }
         repository.delete(autor);
     }
 
@@ -50,6 +57,11 @@ public class AutorService {
         if (autor.getId() == null){
             throw new IllegalArgumentException("Autor não existe");
         }
+        validator.validar(autor);
         repository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
